requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# OpenAI SDK
openai>=1.50.0

# Data validation and parsing
pydantic==2.9.2
pydantic-settings==2.6.0

# File handling
python-multipart==0.0.12
Pillow==10.4.0

# Environment variables
python-dotenv==1.0.1
